chore: consolidate vitest setup
