+++ packages/mvc/vitest.config.ts
@@ -1,23 +1,3 @@
-import { defineConfig } from 'vitest/config';
+import { rootConfig } from '../../vitest.config';
 
-export default defineConfig({
-  test: {
-    setupFiles: ['../../vitest.setup.ts'],
-    coverage: {
-      provider: 'v8',
-      include: ['src/**'],
-      exclude: ['src/index.ts', 'dist'],
-      thresholds: {
-        branches: 100,
-        functions: 100,
-        lines: 100,
-        statements: 100
-      }
-    }
-  },
-  resolve: {
-    alias: {
-      '@expressive/mvc': './src'
-    }
-  }
-});
+export default rootConfig;
