+++ packages/preact/vitest.ts
@@ -1,52 +1,9 @@
+export { mockError, mockPromise, mockWarn } from '../../vitest.setup';
+
 import { cleanup } from '@testing-library/preact';
-import { afterEach, afterAll, vi } from 'vitest';
+import { afterEach } from 'vitest';
 
-afterEach(() => cleanup());
-
-interface CustomMatchers<R = unknown> {
-  /** Assert state does have one or more updates pending. */
-  toUpdate(timeout?: number): Promise<R>;
-
-  /** Assert state did update with keys specified. */
-  toHaveUpdated(...keys: (string | symbol | number)[]): Promise<R>;
-}
-
-declare module 'vitest' {
-  interface Assertion<T = any> extends CustomMatchers<T> {}
-  interface AsymmetricMatchersContaining extends CustomMatchers {}
-}
-
-export interface MockPromise<T> extends Promise<T> {
-  resolve: (value: T | PromiseLike<T>) => void;
-  reject: (reason?: any) => void;
-}
-
-export function mockPromise<T = void>() {
-  const methods = {} as MockPromise<T>;
-  const promise = new Promise((res, rej) => {
-    methods.resolve = res;
-    methods.reject = rej;
-  }) as MockPromise<T>;
-
-  return Object.assign(promise, methods);
-}
-
-export function mockWarn() {
-  const warn = vi.spyOn(console, 'warn').mockImplementation(() => {});
-
-  afterEach(() => warn.mockClear());
-  afterAll(() => warn.mockRestore());
-
-  return warn;
-}
-
-export function mockError() {
-  const error = vi.spyOn(console, 'error').mockImplementation(() => {});
-
-  afterEach(() => error.mockClear());
-  afterAll(() => error.mockRestore());
-
-  return error;
-}
+afterEach(cleanup);
 
 export * from 'vitest';
+export * from '@testing-library/preact';
