+++ vitest.setup.ts
@@ -1,4 +1,4 @@
-import { expect } from 'vitest';
+import { expect, afterEach, afterAll, vi } from 'vitest';
 import { State } from './packages/mvc/src';
 
 // Add @testing-library matchers for React tests
@@ -21,6 +21,8 @@ expect.extend({
   toUpdate,
   toHaveUpdated
 });
+
+export { mockError, mockPromise, mockWarn, MockPromise };
 
 async function toUpdate(received: State, timeout = 0): Promise<any> {
   if (!(received instanceof State))
@@ -90,4 +92,37 @@ async function toHaveUpdated(received: State, ...keys: string[]) {
         .map(String)
         .join(', ')}].`
   };
+}
+
+interface MockPromise<T> extends Promise<T> {
+  resolve: (value: T | PromiseLike<T>) => void;
+  reject: (reason?: any) => void;
+}
+
+function mockPromise<T = void>() {
+  const methods = {} as MockPromise<T>;
+  const promise = new Promise((res, rej) => {
+    methods.resolve = res;
+    methods.reject = rej;
+  }) as MockPromise<T>;
+
+  return Object.assign(promise, methods);
+}
+
+function mockWarn() {
+  const warn = vi.spyOn(console, 'warn').mockImplementation(() => {});
+
+  afterEach(() => warn.mockClear());
+  afterAll(() => warn.mockRestore());
+
+  return warn;
+}
+
+function mockError() {
+  const error = vi.spyOn(console, 'error').mockImplementation(() => {});
+
+  afterEach(() => error.mockClear());
+  afterAll(() => error.mockRestore());
+
+  return error;
 }
