--- vitest.workspace.ts
@@ -1,1 +1,0 @@
-export default ['packages/mvc', 'packages/react'];
